+++ modules/perception/common/graph/connected_component_analysis.h
@@ -34,7 +34,7 @@ namespace common {
 void ConnectedComponentAnalysis(const std::vector<std::vector<int>>& graph,
                                 std::vector<std::vector<int>>* components) {
   CHECK_NOTNULL(components);
-  int num_item = graph.size();
+  int num_item = static_cast<int>(graph.size());
   std::vector<int> visited;
   visited.resize(num_item, 0);
   std::queue<int> que;
